multiple, movable hotspots
OK, going overboard with the hotspots maybe...
The problem with LPs being hotspots is that
they quickly get overloaded with messages, especially on the cluster,
where the MPI incoming message buffer was overflowing and causing
simulations to crash.
Now, multiple simulation objects can be selected as hotspots at different
virtual times.  This makes PHOLD extremely flexible in terms of
the kinds of imbalances we can impose.