Fixed case where sim would fail to terminate
This is a fix for a race condition that only happens when an application
with no internal termination condition is run using non-threaded warped
on multiple nodes with a maximum GVT  specified (i.e. `--simulateUntil
x`). The first manager would terminate as expected, but all other
managers would continue to run.

The manager with ID 0 is hard coded to do GVT calculations. When the max
GVT is exceeded, it would tell the TerminationManager to circulate a
SIMULATION_COMPLETE TerminationToken, which would cause all other
managers to terminate once they received it.

The first manager is also initially in control of triggering the normal
termination check. However, if a manager trigger a normal termination
check (which happens when it runs out of event to process), and another
manager has events remaining, the manager that ran out of events will
hand control of termination over to the manager that had events
remaining.

The first manager can check if it has any events remaining before any
events have been scheduled, causing the termination control to pass to
another manager before the simulation actually begins.

If the first manager had control of the termination, it would trigger
the termination as part of the GVT calculation, which happens even if
there were still pending events. Since only the first manager performs
GVT calculations, if another manager controlled termination, it would
trigger it once it had run out of events to process.

However, it is possible for a simulation object to schedule events for
itself. In the case of models like Epidemic and RAID, all objects will
schedule events for themselves indefinitely.

What was happening was the first manager would calculate the GVT, see
that it had exceeded the max, and trigger a termination. However, if it
had handed off termination control, the termination trigger would be a
no-op, and the other managers would continue to run until the one in
control of termination ran out of events and had a chance to check the
GVT. If that manager never ran out of events, it would never terminate.

The fix was simply to continue to let the first manager perform GVT
calculations, but to have all managers check the GVT value before
processing every event.